Flytt sticky header til egen fil
