+++ nablagreier/lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'top_image_section.dart'; 
-
+import 'sticky_header.dart';
 
 void main() {
   runApp(MyWebPage());
@@ -13,32 +13,13 @@ class MyWebPage extends StatelessWidget {
       home: Scaffold(
         body: CustomScrollView(
           slivers: <Widget>[
-            // This is the top image that will scroll away.
+            
             SliverToBoxAdapter(
-              child: TopImageSection(),
+              child: TopImageSection(),         // This is the top image that will scroll away.
             ),
-            // Sticky header that remains visible as you scroll
-            SliverPersistentHeader(
-              delegate: _StickyHeaderDelegate(
-                minHeight: 60.0,
-                maxHeight: 60.0,
-                child: Container(
-    
-                  color: Color(0xFF1045A6), // Background color of the header              
-                  child: Center(
-                    child: Text(
-                      'Sticky Header',
-                      style: TextStyle(
-                        fontSize: 20,
-                        color: Colors.white,
-                      ),
-                    ),
-                  ),
-                ),
-              ),
-              pinned: true,
-            ),
-            // Your main content goes here, below the sticky header
+            
+            StickyHeader(),                     // Sticky header that remains visible as you scroll
+            
             SliverList(
               delegate: SliverChildBuilderDelegate(
                 (context, index) => ListTile(
@@ -52,35 +33,5 @@ class MyWebPage extends StatelessWidget {
         ),
       ),
     );
-  }
-}
-
-class _StickyHeaderDelegate extends SliverPersistentHeaderDelegate {
-  final double minHeight;
-  final double maxHeight;
-  final Widget child;
-
-  _StickyHeaderDelegate({
-    required this.minHeight,
-    required this.maxHeight,
-    required this.child,
-  });
-
-  @override
-  double get minExtent => minHeight;
-
-  @override
-  double get maxExtent => maxHeight;
-
-  @override
-  Widget build(BuildContext context, double shrinkOffset, bool overlapsContent) {
-    return SizedBox.expand(child: child);
-  }
-
-  @override
-  bool shouldRebuild(_StickyHeaderDelegate oldDelegate) {
-    return minHeight != oldDelegate.minHeight ||
-           maxHeight != oldDelegate.maxHeight ||
-           child != oldDelegate.child;
   }
 }
