+++ nablagreier/lib/sticky_header.dart
@@ -1,0 +1,55 @@
+// sticky_header.dart
+import 'package:flutter/material.dart';
+
+class StickyHeader extends StatelessWidget {
+  const StickyHeader({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return SliverPersistentHeader(
+      delegate: _StickyHeaderDelegate(
+        minHeight: 60.0,
+        maxHeight: 60.0,
+        child: Container(
+          color: Color(0xFF1045A6), // Background color of the header
+          child: Center(
+            child: Text(
+              'Sticky Header',
+              style: TextStyle(fontSize: 20, color: Colors.white),
+            ),
+          ),
+        ),
+      ),
+      pinned: true,
+    );
+  }
+}
+
+class _StickyHeaderDelegate extends SliverPersistentHeaderDelegate {
+  final double minHeight;
+  final double maxHeight;
+  final Widget child;
+
+  _StickyHeaderDelegate({
+    required this.minHeight,
+    required this.maxHeight,
+    required this.child,
+  });
+
+  @override
+  double get minExtent => minHeight;
+  @override
+  double get maxExtent => maxHeight;
+
+  @override
+  Widget build(BuildContext context, double shrinkOffset, bool overlapsContent) {
+    return SizedBox.expand(child: child);
+  }
+
+  @override
+  bool shouldRebuild(_StickyHeaderDelegate oldDelegate) {
+    return minHeight != oldDelegate.minHeight ||
+           maxHeight != oldDelegate.maxHeight ||
+           child != oldDelegate.child;
+  }
+}
